miscを再分類
