+++ src/types/mod.rs
@@ -1,6 +1,11 @@
 pub mod app_state;
 pub mod misc;
 pub mod bot;
+pub mod game;
+pub mod message;
 
 pub use app_state::AppState;
+pub use bot::BotToken;
 pub use misc::*;
+pub use game::*;
+pub use message::*;
